ege-prediction-ml-system/requirements.txt
catboost>=1.2
fastapi>=0.100
uvicorn>=0.23
pydantic>=2.0
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
psycopg2-binary>=2.9
streamlit>=1.28
requests>=2.31
mlflow==2.16.0
